Return error when parsing request meant for other slave
`modbusParseRequestRTU()` now returns `MODBUS_REQUEST_ERROR(ADDRESS)`
if the provided frame is meant for other slave. Since request errors
are usually ignored on slave side, this shouldn't be too much of a
problem. `modbusIsOk()` returning false ensures that no attempt is
made to respond to the master.

Silently ignoring frames meant for other slaves may be a not very
good idea, as pointed out in #15. This change makes it possible for the
user to know exactly whether the slave took any action, and if not, for
what reason. `modbusIsOk()` returning true now means that slave
has successfully processed the frame and has executed requested function.