Return `MODBUS_RESPONSE_ERROR(LENGTH)` if a response is generated to a broadcast RTU request
If the handler function creates a response even though the request is
broadcasted, report MODBUS_RESPONSE_ERROR(LENGTH). Reporting a
general error here may be a bad idea, because it could cause
quite serious consequences in the user applicaton, while the error
is in fact very easy to handle internally.

This approach also allows users to write custom functions
that always generate responses and rely on this mechanism
in order to avoid sending responses to broadcast requests.
Doing so, however, is not recommended.

Importantly, returning this error results in `modbusIsOk()`
evaluating to false, ensuring that the response would
not be sent.